+++ spring-mvc-basics/src/main/java/com/baeldung/customvalidator/ContactNumberValidator.java
@@ -1,0 +1,17 @@
+package com.baeldung.customvalidator;
+
+import javax.validation.ConstraintValidator;
+import javax.validation.ConstraintValidatorContext;
+
+public class ContactNumberValidator implements ConstraintValidator<ContactNumberConstraint, String> {
+
+    @Override
+    public void initialize(ContactNumberConstraint contactNumber) {
+    }
+
+    @Override
+    public boolean isValid(String contactField, ConstraintValidatorContext cxt) {
+        return contactField != null && contactField.matches("[0-9]+") && (contactField.length() > 8) && (contactField.length() < 14);
+    }
+
+}
