+++ spring-mvc-basics/src/main/java/com/baeldung/web/controller/ValidatedPhoneController.java
@@ -1,0 +1,32 @@
+package com.baeldung.web.controller;
+
+import javax.validation.Valid;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.validation.BindingResult;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+
+import com.baeldung.model.ValidatedPhone;
+
+@Controller
+public class ValidatedPhoneController {
+
+    @GetMapping("/validatePhone")
+    public String loadFormPage(Model m) {
+        m.addAttribute("validatedPhone", new ValidatedPhone());
+        return "phoneHome";
+    }
+
+    @PostMapping("/addValidatePhone")
+    public String submitForm(@Valid ValidatedPhone validatedPhone, BindingResult result, Model m) {
+        if (result.hasErrors()) {
+            return "phoneHome";
+        }
+
+        m.addAttribute("message", "Successfully saved phone: " + validatedPhone.toString());
+        return "phoneHome";
+    }
+
+}
