--- spring-mvc-java/src/main/java/com/baeldung/customvalidator/ContactNumberConstraint.java
@@ -1,24 +1,0 @@
-package com.baeldung.customvalidator;
-
-import java.lang.annotation.Documented;
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-import javax.validation.Constraint;
-import javax.validation.Payload;
-
-@Documented
-@Constraint(validatedBy = ContactNumberValidator.class)
-@Target({ElementType.METHOD, ElementType.FIELD})
-@Retention(RetentionPolicy.RUNTIME)
-public @interface ContactNumberConstraint {
-
-    String message() default "Invalid phone number";
-
-    Class<?>[] groups() default {};
-
-    Class<? extends Payload>[] payload() default {};
-
-}
