--- spring-mvc-java/src/main/java/com/baeldung/model/ValidatedPhone.java
@@ -1,22 +1,0 @@
-package com.baeldung.model;
-
-import com.baeldung.customvalidator.ContactNumberConstraint;
-
-public class ValidatedPhone {
-
-    @ContactNumberConstraint
-    private String phone;
-
-    public String getPhone() {
-        return phone;
-    }
-
-    public void setPhone(String phone) {
-        this.phone = phone;
-    }
-
-    @Override
-    public String toString() {
-        return phone;
-    }
-}
